+++ Assets/Scripts/Networking/NetworkPlayerWeapon.cs
@@ -5,6 +5,6 @@
 public class NetworkPlayerWeapon {
 
 	public string name = "Raygun";
-	public int damage = 25;
+	public float damage = 25f;
 	public float range = 50f;
 }
